Add support for gorilla/mux...
Subrouters and HTTP methods.

Prior to this change, when a subrouter was created, it would result in a
panic when trying to get thee function name.  Now, when a Subrouter is encountered, it will simply print
"SUBROUTER" for nte function name.

Additionally, if HTTP Methods are set for a route, it will now extract
the supported methods, "stringify" them, and add them to the display.